+++ Sources/MapboxMaps/Annotations/AnnotationOrchestrator.swift
@@ -18,14 +18,14 @@ public protocol AnnotationManager: AnyObject {
     var slot: String? { get set }
 }
 
-internal protocol AnnotationManagerInternal: AnnotationManager {
+protocol AnnotationManagerInternal: AnnotationManager {
     var allLayerIds: [String] { get }
 
     func destroy()
 
-    func handleTap(with featureId: String, context: MapContentGestureContext) -> Bool
+    func handleTap(layerId: String, feature: Feature, context: MapContentGestureContext) -> Bool
 
-    func handleLongPress(with featureId: String, context: MapContentGestureContext) -> Bool
+    func handleLongPress(layerId: String, feature: Feature, context: MapContentGestureContext) -> Bool
 
     func handleDragBegin(with featureId: String, context: MapContentGestureContext) -> Bool
 
@@ -57,7 +57,7 @@ public protocol AnnotationInteractionDelegate: AnyObject {
 public final class AnnotationOrchestrator {
     private let impl: AnnotationOrchestratorImplProtocol
 
-    internal init(impl: AnnotationOrchestratorImplProtocol) {
+    init(impl: AnnotationOrchestratorImplProtocol) {
         self.impl = impl
     }
 
@@ -74,9 +74,13 @@ public final class AnnotationOrchestrator {
     ///   - layerPosition: Optionally set the `LayerPosition` of the layer managed.
     ///   - clusterOptions: Optionally set the `ClusterOptions` to cluster the Point Annotations
     /// - Returns: An instance of `PointAnnotationManager`
-    public func makePointAnnotationManager(id: String = String(UUID().uuidString.prefix(5)),
-                                           layerPosition: LayerPosition? = nil,
-                                           clusterOptions: ClusterOptions? = nil) -> PointAnnotationManager {
+    public func makePointAnnotationManager(
+        id: String = String(UUID().uuidString.prefix(5)),
+        layerPosition: LayerPosition? = nil,
+        clusterOptions: ClusterOptions? = nil,
+        onClusterTap: ((AnnotationClusterGestureContext) -> Void)? = nil,
+        onClusterLongPress: ((AnnotationClusterGestureContext) -> Void)? = nil
+    ) -> PointAnnotationManager {
         // swiftlint:disable:next force_cast
         return impl.makePointAnnotationManager(id: id, layerPosition: layerPosition, clusterOptions: clusterOptions) as! PointAnnotationManager
     }